Add EndReceiver to CkSyncBarrier
This allows users to add a callback to run after the normal AtSync
calbacks have finished, which can be used for cleanup or to prepare
for resuming normal execution. Namely, this is used by CkLocMgr in
order to tell the LB framework that it has stopped inserting objects,
which was previously done by shoehorning it into being an AtSync
client unnaturally.

Additionally, the flooding mechanism of AtSync, intended to cover
cases where there are empty PEs, can hang when the flood message is
received by a PE before that PE has exited its AtSync section, meaning
that its CkSyncBarrier is off. To fix this, track the epoch of the
flood message (now called "kick" to more closely match the semantics
of the procedure) so the barrier can check if the kick for the current
epoch has been received when the barrier is eventually enabled again.